# JAMES-2448 Enhance jmap memory tests