--- a/pom.xml
+++ b/pom.xml
@@ -582,7 +582,7 @@
 
         <activemq.version>5.15.2</activemq.version>
         <apache-mime4j.version>0.8.2</apache-mime4j.version>
-        <camel.version>2.19.4</camel.version>
+        <camel.version>2.21.1</camel.version>
         <derby.version>10.9.1.0</derby.version>
         <hadoop.version>1.1.1</hadoop.version>
         <hbase.version>0.94.27</hbase.version>