--- a/server/mailet/mailetcontainer-camel/src/main/java/org/apache/james/mailetcontainer/impl/MatcherMailetPair.java
+++ b/server/mailet/mailetcontainer-camel/src/main/java/org/apache/james/mailetcontainer/impl/MatcherMailetPair.java
@@ -51,4 +51,9 @@ public class MatcherMailetPair {
         return mailet;
     }
 
+    public String getOnMatchException() {
+        return mailet.getMailetConfig()
+            .getInitParameter("onMatchException");
+    }
+
 }