… server/protocols/jmap-integration-testing/jmap-integration-testing-common/src/test/java/org/apache/james/jmap/TestingConstants.java
… server/protocols/jmap-integration-testing/memory-jmap-integration-testing/src/test/resources/imapserver.xml
… server/protocols/jmap-integration-testing/memory-jmap-integration-testing/src/test/resources/lmtpserver.xml
… server/protocols/jmap-integration-testing/memory-jmap-integration-testing/src/test/resources/managesieveserver.xml
… server/protocols/jmap-integration-testing/memory-jmap-integration-testing/src/test/resources/pop3server.xml
… server/protocols/jmap-integration-testing/memory-jmap-integration-testing/src/test/resources/smtpserver.xml