Before: 1 test = 2'' -> 2''400

Now: 1 test = 300ms -> 400ms

Actions taken:
 - Calmly await should get read of poll delay, and try directly (~1s gain)
 - Simplify logic pushed to camel, don't push inefficient choices. With other camel enhancements, this brings init time from 1s to ~200ms.

Performance tests demonstrating the impact on SMTP performance will  soon follow (I'm also expecting a gain).