@chibenwa 2018-07-02
SMTP delivery performance are equivalent before and after this PR.

I run 4 gatling tests on my machine on top of memory-guice.

Before:

100 users, 97.557 mail/s, 99 percentile at 9 ms

1000 users, 785.271 mail/s, 99 percentile at 29ms

After:

100 users, 97.556 mail/s, 99 percentile at 9 ms

1000 users, 785.819 mail/s, 99 percentile at 30 ms