@rouazana 2018-07-02
```
[2a274cd170cbeb1f807753a0b07d4a265e28d818] [INFO] Apache James :: Server :: JMAP :: Cassandra Integration testing SUCCESS [50:00 min]
[2a274cd170cbeb1f807753a0b07d4a265e28d818] [INFO] Apache James :: Server :: JMAP :: Memory Integration testing SUCCESS [07:46 min]
```