@chibenwa 2018-07-02
We should also include mailet integration testing in the stat...