@chibenwa 2018-07-03
Was :green_apple: , I rebased, remove reverted commits, & reword to add ticket numbers.