@chibenwa 2018-07-04
```
[2ee0e045187219ecb3c3c96368ff26c5268f16ea] Tests in error: 
[2ee0e045187219ecb3c3c96368ff26c5268f16ea]   DirectResolutionRemoteDeliveryIntegrationTest.directResolutionShouldBeWellPerformed:113 » ConditionTimeout
```